Moved rllib_proxy/_trainer_template.py to proxy/agents/trainer_template.py
